Do some cleanup
